Make node size always 64 bits
This ensures that sizeof(SerdNode) is a power of two, even on 32-bit platforms,
which is required by posix_memalign.

+++ src/node.h
@@ -22,7 +22,7 @@
 #include <stddef.h>
 
 struct SerdNodeImpl {
-	size_t        n_bytes;  /**< Size in bytes (not including null) */
+	uint64_t      n_bytes;  /**< Size in bytes (not including null) */
 	SerdNodeFlags flags;    /**< Node flags (e.g. string properties) */
 	SerdNodeType  type;     /**< Node type */
 };
